修复了被动模式的bug，但是现在有时刚开始的时候还不行

+++ ihealthEx/passive_mode.h
@@ -4,7 +4,7 @@
 
 /*被动动作参数，记录被动运动的起点和终点的位置、速度和插值点数*/
 struct PassiveActionParam {
-	PassiveActionParam(double shoulder_start_pos, double shoulder_end_pos, double elbow_start_pos, double elbow_end_pos, int step) :
+	PassiveActionParam(double shoulder_start_pos, double shoulder_end_pos, double elbow_start_pos, double elbow_end_pos, double step) :
 		shoulder_start_position(shoulder_start_pos), shoulder_end_position(shoulder_end_pos), elbow_start_position(elbow_start_pos),
 		elbow_end_position(elbow_end_pos), total_step(step) {
 	
@@ -20,7 +20,7 @@ struct PassiveActionParam {
 	double elbow_start_velocity = 0;
 	double elbow_end_velocity = 0;
 
-	int total_step;
+	double total_step;
 };
 
 /**/
@@ -39,6 +39,6 @@ public:
 private:
 	std::shared_ptr<ControlCard> p_control_card_;
 	void PositionMove(const  PassivePosition &position);
-	PassivePosition CalculateAxisPosition(const PassiveActionParam &param, int step);
+	PassivePosition CalculateAxisPosition(const PassiveActionParam &param, double step);
 	void WaitTillMoveFinish();
 };
